feat(auth): implement resend OTP functionality for forgot password
add the resend otp for forgot password and changed the logic in the resend otp to work with both signup and forgot password.
change email saving variable from req.session.userDetails to the req.session.email in signup to adapt common for the resend function.
change forgotOtp to otp and forgotOtpExpiry to otpExpiry in forgotPassword logic for making code similar.